backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.5
python-dotenv==1.0.0
moviepy>=1.0.3

